+++ CHANGELOG.md
@@ -4,6 +4,10 @@
  * Init of project.
  * Add base makefile compilation.
  * Unit tests.
+ * Base library:
+  * Assert,
+  * String number parser,
+  * Type define.
 
 ## Fixed